+++ src/telegram/mod.rs
@@ -8,6 +8,7 @@ use std::path::PathBuf;
 
 use crate::config::DailyScryConfig;
 use crate::error::Result;
+use crate::util::{split_text, Additional};
 
 use teloxide_core::{
     payloads::{SendMessageSetters, SendPhotoSetters},
@@ -30,7 +31,15 @@ pub async fn post(
     let chat_id = config.telegram_chat_id.clone().unwrap();
 
     let futures = images_and_texts.map(|(image, card_text)| {
-        map_function(&bot, &chat_id, artist.clone(), &image, card_text, link)
+        map_function(
+            &bot,
+            &chat_id,
+            artist.clone(),
+            &image,
+            card_text,
+            link,
+            config,
+        )
     });
 
     futures::future::join_all(futures)
@@ -47,9 +56,18 @@ async fn map_function(
     image_path: &PathBuf,
     text: &String,
     link: &str,
+    config: &DailyScryConfig,
 ) -> Result<()> {
     send_image(bot, chat_id, image_path, link).await?;
-    send_message(bot, chat_id, artist, text).await?;
+    let artist = artist.unwrap_or_default();
+    let splitted_texts = split_text(
+        text.to_string(),
+        config.telegram_character_limit.unwrap(),
+        vec![Additional::Text(artist.clone())],
+    );
+    for text in splitted_texts {
+        send_message(bot, chat_id, &artist, text).await?;
+    }
     Ok(())
 }
 
@@ -61,13 +79,8 @@ async fn send_image(bot: &Bot, chat_id: &String, image_path: &PathBuf, link: &st
     Ok(())
 }
 
-async fn send_message(
-    bot: &Bot,
-    chat_id: &String,
-    artist: Option<String>,
-    text: &String,
-) -> Result<()> {
-    let text_with_artist = format!("{}{}", text, artist.unwrap_or_default());
+async fn send_message(bot: &Bot, chat_id: &String, artist: &String, text: String) -> Result<()> {
+    let text_with_artist = format!("{}{}", text, artist);
     bot.send_message(chat_id.clone(), text_with_artist)
         .parse_mode(ParseMode::Html)
         .send()